+++ src/pos/earth.rs
@@ -6,9 +6,8 @@
 
 use std::fmt::Display;
 
-use erfa_sys::{ERFA_GRS80, ERFA_WGS72, ERFA_WGS84};
+use erfa::Ellipsoid;
 
-use super::ErfaError;
 use crate::{
     constants::{MWA_HEIGHT_M, MWA_LAT_RAD, MWA_LONG_RAD},
     XyzGeocentric,
@@ -23,16 +22,6 @@ pub struct LatLngHeight {
     pub latitude_rad: f64,
     /// Height above ellipsoid \[meters\]
     pub height_metres: f64,
-}
-
-/// Enum of erfa-compatible reference ellipsoids.
-pub enum Ellipsoid {
-    /// WGS84 reference ellipsoid
-    WGS84 = ERFA_WGS84 as isize,
-    /// GRS80 reference ellipsoid
-    GRS80 = ERFA_GRS80 as isize,
-    /// WGS72 reference ellipsoid
-    WGS72 = ERFA_WGS72 as isize,
 }
 
 impl LatLngHeight {
@@ -51,44 +40,26 @@ impl LatLngHeight {
         Self::mwa()
     }
 
-    /// Convert to [`XyzGeocentric`] via [`erfa_sys::eraGd2gc`] with the specified
+    /// Convert to [`XyzGeocentric`] via
+    /// [`erfa::transform::geodetic_to_geocentric`] with the specified
     /// [`Ellipsoid`]
-    ///
-    /// # Errors
-    ///
-    /// Can return an [`ErfaError`] if [`erfa_sys::eraGd2gc`] fails.
-    pub fn to_geocentric(self, ellipsoid: Ellipsoid) -> Result<XyzGeocentric, ErfaError> {
-        let mut geocentric_vector: [f64; 3] = [0.0; 3];
-        let status = unsafe {
-            erfa_sys::eraGd2gc(
-                ellipsoid as i32,               // ellipsoid identifier (Note 1)
-                self.longitude_rad,             // longitude (radians, east +ve)
-                self.latitude_rad,              // latitude (geodetic, radians, Note 3)
-                self.height_metres,             // height above ellipsoid (geodetic, Notes 2,3)
-                geocentric_vector.as_mut_ptr(), // geocentric vector (Note 2)
-            )
-        };
-        if status != 0 {
-            return Err(ErfaError {
-                source_file: file!(),
-                source_line: line!(),
-                status,
-                function: "eraGd2gc",
-            });
-        }
-        Ok(XyzGeocentric {
+    pub fn to_geocentric(self, ellipsoid: Ellipsoid) -> XyzGeocentric {
+        let geocentric_vector = erfa::transform::geodetic_to_geocentric(
+            ellipsoid,
+            self.longitude_rad,
+            self.latitude_rad,
+            self.height_metres,
+        )
+        .expect("latitude should be between -pi/2 and pi/2");
+        XyzGeocentric {
             x: geocentric_vector[0],
             y: geocentric_vector[1],
             z: geocentric_vector[2],
-        })
+        }
     }
 
     /// Convert to geocentric via the default [`Ellipsoid::WGS84`].
-    ///
-    /// # Errors
-    ///
-    /// Can return an [`ErfaError`] if [`erfa_sys::eraGd2gc`] fails.
-    pub fn to_geocentric_wgs84(self) -> Result<XyzGeocentric, ErfaError> {
+    pub fn to_geocentric_wgs84(self) -> XyzGeocentric {
         self.to_geocentric(Ellipsoid::WGS84)
     }
 }